Convert forceVisualforcePageCreate
