--- packages/salesforcedx-vscode-core/src/commands/forceVisualforcePageCreate.ts
@@ -1,111 +1,0 @@
-/*
- * Copyright (c) 2017, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-
-import {
-  CliCommandExecutor,
-  Command,
-  SfdxCommandBuilder
-} from '@salesforce/salesforcedx-utils-vscode/out/src/cli';
-import {
-  ContinueResponse,
-  DirFileNameSelection
-} from '@salesforce/salesforcedx-utils-vscode/out/src/types';
-import * as path from 'path';
-import { Observable } from 'rxjs/Observable';
-import * as vscode from 'vscode';
-import { channelService } from '../channels';
-import { nls } from '../messages';
-import { notificationService, ProgressNotification } from '../notifications';
-import { taskViewService } from '../statuses';
-import {
-  CompositeParametersGatherer,
-  FilePathExistsChecker,
-  SelectFileName,
-  SelectOutputDir,
-  SfdxCommandlet,
-  SfdxCommandletExecutor,
-  SfdxWorkspaceChecker
-} from './commands';
-
-const VF_PAGE_EXTENSION = '.page';
-const VF_PAGE_METADATA_DIR = 'pages';
-
-class ForceVisualForcePageCreateExecutor extends SfdxCommandletExecutor<
-  DirFileNameSelection
-> {
-  public build(data: DirFileNameSelection): Command {
-    return new SfdxCommandBuilder()
-      .withDescription(nls.localize('force_visualforce_page_create_text'))
-      .withArg('force:visualforce:page:create')
-      .withFlag('--pagename', data.fileName)
-      .withFlag('--label', data.fileName)
-      .withFlag('--outputdir', data.outputdir)
-      .withLogName('force_visualforce_page_create')
-      .build();
-  }
-
-  public execute(response: ContinueResponse<DirFileNameSelection>): void {
-    const startTime = process.hrtime();
-    const cancellationTokenSource = new vscode.CancellationTokenSource();
-    const cancellationToken = cancellationTokenSource.token;
-
-    const execution = new CliCommandExecutor(this.build(response.data), {
-      cwd: vscode.workspace.rootPath
-    }).execute(cancellationToken);
-
-    execution.processExitSubject.subscribe(async data => {
-      const dirType = response.data.outputdir.endsWith(
-        path.join(SelectOutputDir.defaultOutput, VF_PAGE_METADATA_DIR)
-      )
-        ? 'defaultDir'
-        : 'customDir';
-      this.logMetric(execution.command.logName, startTime, { dirType });
-      if (
-        data !== undefined &&
-        data.toString() === '0' &&
-        vscode.workspace.rootPath
-      ) {
-        vscode.workspace
-          .openTextDocument(
-            path.join(
-              vscode.workspace.rootPath,
-              response.data.outputdir,
-              response.data.fileName + VF_PAGE_EXTENSION
-            )
-          )
-          .then(document => vscode.window.showTextDocument(document));
-      }
-    });
-
-    notificationService.reportExecutionError(
-      execution.command.toString(),
-      (execution.stderrSubject as any) as Observable<Error | undefined>
-    );
-
-    channelService.streamCommandOutput(execution);
-    ProgressNotification.show(execution, cancellationTokenSource);
-    taskViewService.addCommandExecution(execution, cancellationTokenSource);
-  }
-}
-
-const workspaceChecker = new SfdxWorkspaceChecker();
-const fileNameGatherer = new SelectFileName();
-const filePathExistsChecker = new FilePathExistsChecker(VF_PAGE_EXTENSION);
-
-export async function forceVisualforcePageCreate() {
-  const outputDirGatherer = new SelectOutputDir(VF_PAGE_METADATA_DIR);
-  const parameterGatherer = new CompositeParametersGatherer<
-    DirFileNameSelection
-  >(fileNameGatherer, outputDirGatherer);
-  const commandlet = new SfdxCommandlet(
-    workspaceChecker,
-    parameterGatherer,
-    new ForceVisualForcePageCreateExecutor(),
-    filePathExistsChecker
-  );
-  await commandlet.run();
-}
